Rename info circle and sign out SVGs
