+++ frontend/assets/svg/ic_info_circle_24dp.js
