+++ frontend/assets/svg/ic_sign_out_24dp.js
