+++ frontend/assets/svg/index.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import PlaceholderIcon from './ic_placeholder_24dp'
-import InfoIcon from './ic_info_24dp'
+import InfoCircleIcon from './ic_info_circle_24dp'
 import CrosshairsIcon from './ic_crosshairs_24dp'
 import MegaphoneIcon from './ic_megaphone_24dp'
-import SignOutIcon from './ic_signout_24dp'
+import SignOutIcon from './ic_sign_out_24dp'
 import TwitterIcon from './ic_twitter_24dp'
 
 
@@ -12,7 +12,7 @@ const Icon = props => {
     case "crosshairs":
       return <CrosshairsIcon {...props} />;
     case "info":
-      return <InfoIcon {...props} />;
+      return <InfoCircleIcon {...props} />;
     case "megaphone":
       return <MegaphoneIcon {...props} />;
     case "signout":
